Ensures device consistency for tensors
Addresses a issue where tensors were not consistently placed on the correct device (CPU or GPU). Also made sure all types are correct and consistent

Updates the code to explicitly move tensors to the appropriate device, ensuring compatibility and preventing runtime errors.

Specifically, uses the `.to(atom_list.device)` method instead of relying on default tensor placement.
Also, casts numpy arrays to float32 dtype.

+++ code/AttentiveFP/Featurizer_aromaticity_rm.py
@@ -67,7 +67,7 @@ def atom_features(atom,
                 results = results + [False, False
                                      ] + [atom.HasProp('_ChiralityPossible')]
 
-        return np.array(results)
+        return np.array(results, dtype=np.float32)
 
 
 def bond_features(bond, use_chirality=True):
@@ -82,7 +82,7 @@ def bond_features(bond, use_chirality=True):
         bond_feats = bond_feats + one_of_k_encoding_unk(
             str(bond.GetStereo()),
             ["STEREONONE", "STEREOANY", "STEREOZ", "STEREOE"])
-    return np.zeros(shape=np.array(bond_feats).shape) # set all bond features blank
+    return np.zeros(shape=np.array(bond_feats).shape, dtype=np.float32) # set all bond features blank
 
 
 def num_atom_features():
